requirements.txt
together==0.2.5
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
python-docx==1.0.1
PyPDF2==3.0.1
Pillow==10.1.0
openpyxl==3.1.2
python-dotenv==1.0.0 